fix: isolate idle handler tests in subprocess to avoid mock pollution
Run idle handler tests in a separate Bun process so mock.module
calls on @loreai/core don't leak into cache-warmer and other tests
that import the same module.
