+++ .lore.md
@@ -28,7 +28,7 @@
 * **Bun tests leak into production DB when run outside repo root**: Running \`bun test\` outside the repo means \`bunfig.toml\` is not found, preload never fires, and tests silently open the production DB. Fix (now merged to main): \`db()\` throws immediately if \`NODE\_ENV=test\` and \`LORE\_DB\_PATH\` is not set to a temp path — runtime guard rather than relying on preload. \`bunfig.toml\`-based isolation is fragile due to CWD dependency.
 
 <!-- lore:019e1d28-e23e-7a64-beb9-fa617fac79cd -->
-* **cachedProjectPath in pipeline.ts causes cross-project misattribution for compaction/idle work**: In \`packages/gateway/src/pipeline.ts\`, \`cachedProjectPath\` froze the first-ever resolved project path, causing all subsequent compaction runs to be attributed to it. Similarly, \`buildIdleWorkHandler\` in \`idle.ts\` closed over a single \`projectPath\` parameter captured at handler-build time, ignoring \`state.projectPath\`. Fix (implemented): removed \`cachedProjectPath\` entirely from \`pipeline.ts\` — \`handleCompaction\` now calls \`inferProjectPath()\`/\`getProjectPath()\` fresh each time. Changed \`buildIdleWorkHandler\` signature to remove the \`projectPath\` param; handler now reads \`state.projectPath\` at call time. Call site in \`pipeline.ts\` updated accordingly. The \`process.cwd()\` fallback and greedy generic regex in \`inferProjectPath\` can still produce parent-directory paths like \`/home/byk/Code\` → project name "Code", so always prefer session-derived paths.
+* **cachedProjectPath in pipeline.ts causes cross-project misattribution for compaction/idle work**: In \`packages/gateway/src/pipeline.ts\`, \`cachedProjectPath\` froze the first-ever resolved project path, causing all subsequent compaction runs to be attributed to it. Similarly, \`buildIdleWorkHandler\` in \`idle.ts\` closed over a single \`projectPath\` parameter captured at handler-build time, ignoring \`state.projectPath\`. Fix (implemented): removed \`cachedProjectPath\` entirely from \`pipeline.ts\`; \`handleCompaction\` now calls \`inferProjectPath()\` directly. Changed \`buildIdleWorkHandler\` to remove the \`projectPath\` parameter — reads \`state.projectPath\` at call time instead. The \`process.cwd()\` fallback and greedy generic regex in \`inferProjectPath\` can produce parent-directory paths like \`/home/byk/Code\` → project name "Code" as a remaining risk.
 
 <!-- lore:019e1c27-967f-7ae6-8b22-41332a4ed564 -->
 * **embedding.resetProvider() in test teardown causes worker crash via fastembed NAPI**: Calling \`embedding.resetProvider()\` in \`resetPipelineState()\` (test harness teardown) sets \`cachedProvider = undefined\`, allowing pipeline re-init to spawn new fastembed workers. If those workers are alive at process exit, Bun's NAPI teardown crashes (\`workers\_terminated\` mismatch). Fix: move \`embedding.resetProvider()\` to \`shutdown()\` in \`start.ts\` (called only at real process exit via \`safeExit()\`). \`resetPipelineState()\` must NOT call \`resetProvider()\` — preserve the 'fastembed unavailable' cached state so test re-runs don't re-spawn workers.
